requirements.txt
Flask==2.2.5
gunicorn==21.2.0
requests==2.31.0
urllib3==1.26.6
pymongo==4.5.0
bcrypt==4.0.1
pytz==2023.3.post1
googlesearch-python==1.2.3